Update cloud-init-troubleshooting.md
Removed chroot from "Step3: Collect & review VM logs".  Chroot isn't a requirement to access an oskdisk filesystem on a rescue VM. Using such could make troubleshooting more complicated if you intend on pulling off the logs in the advent of a support case requirement. 

Added simply mount to the section providing a link to the instruction for collecting and/or examining logs both LVM and Raw.

Added cloud-init.log and cloud.init-output.log as these are part of the initial investigation.

Added to "deeper dive" section serial logs as they are very important to look at for provisioning failures. Provided link how to access boot diagnostics (serial log) in the portal.