Update container-instances-quickstart-powershell.md to fix incorrect PowerShell command
The PowerShell command used in the example code was not correct. It also did not have reference to the container instance that was created in the previous code snippet.

I have corrected the powershell cmdlet and also added a new variable for the container instance and that is now referenced in New-AzContainerGroup.